@@ -13,12 +13,14 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.3.3.RELEASE</version>
+    <version>2.5.6</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>11</java.version>
+    <junit-jupiter.version>5.8.0</junit-jupiter.version>
+
   </properties>
 
   <dependencies>
