@@ -1,7 +1,6 @@
 package victor.training.jpa.perf;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +15,6 @@ import javax.persistence.EntityManager;
 
 import static java.lang.System.currentTimeMillis;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 @Transactional
 @Rollback(false)
